Parallelize get_tree_size
This commit replaces the assignment of a subdir crawl thread to
a top-level subdirectory, and instead attempts to assign EVERY
subdir crawl to a new thread.

However, at the same time a thread budget (maxthreads) is imposed,
so when the number of outstanding crawlers is equal to maxthreads,
new directories are crawled recursively in the same crawl thread.

Each iteration of the crawler loop now contains a poll of the
outstanding subdir crawlers to harvest the ones which have completed.
This allows new subdirectories to be crawled in parallel as the
thread budget allows.

This commit also changes the accumulation of ES documents from a
stack-resident parameter ("docs") to a global thread-sharded
map of lists. These are flushed as needed to ES when the chunk
size is reached. This change simplifies the bookkeeping and also
keeps this buffer off the stack, allowing subdir crawls to be
launched and harvested without perturbing this state.